shahkar_sample/requirements.txt
Django>=4.0,<5.0
celery>=5.2,<6.0
redis>=4.0,<5.0
psycopg2-binary>=2.9,<3.0
faker
pyspark
django-celery-results
djangorestframework
drf-spectacular
pandas
eventlet


